// frontend/src/main/frontend/src/app/feature/shared/models/address.model.ts
/*
  export class Address  {
   nos:number;
   address1: string;  
   address2: string; 
   city: string;
   state: string;
    

    constructor(address:any){
            this.address1 = address.address1;
            this.address2 = address.address2;
            this.state = address.State;
            this.city=address.City;
    }
} 
*/
export class Address {
        address1: string;  
        address2: string; 
        city: string;
        state: string;
}